Fix a memory leak. (#2)
shark_output was locally defined, which caused the global shark_output to be left as NULL. This meant that the compiler output would never be destroyed, and resources would be left wasted. The amount of memory lost is at a minimum of ~5KB for a single, simple shader program.

Additionally, sceShaccCgReleaseCompiler will now release resources not freed by DestroyCompileOutput. It is only ~64B.